Lifted state up
Stored state in the Board component instead of in each Square so that the Board component can tell each Square what to display.
Added a constructor to the Board and set its initial state to contain an array with 9 nulls, corresponding to the 9 squares.
Modified renderSqure to pass a value prop to Square.